agent/requirements.txt
pywin32>=306
requests>=2.28.0
pyyaml>=6.0
urllib3>=1.26.0
